{"schema":"decodex/commit/1","summary":"Modularize merge readback and operator boundaries","authority":"manual"}
