+++ apps/decodex/src/orchestrator/operator_http/types/routes.rs
@@ -1,0 +1,21 @@
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub(in crate::orchestrator::operator_http) enum OperatorRequestRoute {
+	Dashboard,
+	DashboardIconPng,
+	DashboardLogoIco,
+	DashboardLogoTouchPng,
+	DashboardWs,
+	Live,
+	AppSnapshot,
+	LinearScan,
+	LaneInspect,
+	LaneInterrupt,
+	LaneSteer,
+	AccountList { force_refresh: bool },
+	AccountSelect,
+	AccountClear,
+	AccountLogout,
+	AccountImport,
+	AccountUse,
+	AccountRerollName,
+}
